+++ repomatic/templates/release-notes.md
@@ -1,5 +1,5 @@
 ---
-args: [version, date, compare_url, development_warning, yanked_admonition, availability_admonition, changes]
+args: [version, date, compare_url, development_warning, yanked_admonition, availability_admonition, editorial_admonition, changes]
 ---
 
 ## [`\$version` (\$date)]($compare_url)
@@ -7,6 +7,8 @@ args: [version, date, compare_url, development_warning, yanked_admonition, avail
 \$development_warning
 
 \$yanked_admonition
+
+\$editorial_admonition
 
 \$availability_admonition